Sit/Chair handler renaming
